bloom and json tests fixed
