+++ redisplus/bf/client.py
@@ -4,7 +4,7 @@ from redis.client import Pipeline, bool_ok
 from redis.commands import Commands as RedisCommands
 
 from .commands import CommandMixin
-from .utils import parseToList
+from ..helpers import parseToList
 from .info import (
     BFInfo,
     CFInfo,
