--- redisplus/bf/utils.py
@@ -1,17 +1,0 @@
-from ..utils import nativestr
-
-
-def spaceHolder(response):
-    """Return the response without parsing."""
-    return response
-
-
-def parseToList(response):
-    """Parse the response to a list."""
-    res = []
-    for item in response:
-        if item is not None:
-            res.append(nativestr(item))
-        else:
-            res.append(None)
-    return res
